Put together the first stable version of the app
The whole purpose of making this website was to challenge my new
React skills, such as the heavy usage of state and props, especially
in sharing the data submitted by the user between the "edit" and
"preview" modes.
I'd like to add that create-react-app was not used in this project
as a way of practicing setting up the React code environment, with
Webpack and Babel.